# server/test/swagger.yaml
swagger: '2.0'
info:
  title: Property Pro Lite
  description: Property Pro Lite is a platform where people can create and / or search properties for sale or rent.
  version: 1.0
servers:
  - url: 'https://pplite-challenge3.herokuapp.com/'
  - url: 'http://localhost:8080'
...
paths:
...
  /api/v2/property:
    get:
      tags:
        - Properties
      summary: Get property by type
      parameters:
        - name: type
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: string
      responses:
        '200':
          description: Properties of Specified Type Are Listed Successful
          content:
            application/json:
              examples:
                foo:
                  value:
                    status: 200
                    data:
                      - price: ' 8000000'
                        state: ' Rwanda'
                        city: ' Kigali'
                        address: ' KK 1 st'
                        type: ' 3 bedroom'
                        owner: 1
                        id: 1
                        status: available
                        created_on: '2019-07-16T00:21:54+02:00'
                        image_url: 'http://res.cloudinary.com/mucyomiller/image/upload/v1563229308/wel7wues5sv8cro8pvzg.png'
        '404':
          description: Resource Is Not Found
          content:
            application/json:
              example:
                status: 404
                error: Route Not Found!
    post:
      tags:
        - Properties
      summary: Create a  property
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProperty'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/body'
        required: true
      responses:
        '201':
          description: Property Created Successful
          content:
            application/json:
              examples:
                foo:
                  value:
                    status: 201
                    data:
                      price: ' 8000000'
                      state: ' Rwanda'
                      city: ' Kigali'
                      address: ' KK 1 st'
                      type: ' 3 bedroom'
                      owner: 1
                      id: 1
                      status: available
                      created_on: '2019-07-16T00:21:54+02:00'
                      image_url: 'http://res.cloudinary.com/mucyomiller/image/upload/v1563229308/wel7wues5sv8cro8pvzg.png'
        '400':
          description: User Supplied Malformatted Data
          content:
            application/json:
              example:
                status: 400
                errors:
                  - address is required
                  - addresss is not allowed
        '401':
          description: Access token is missing or invalid
          content:
            application/json:
              example:
                status: 401
                message: Authentication failed!
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    CreateProperty:
      type: object
      properties:
        price:
          type: integer
          example: 8000000
        state:
          type: string
          example: Rwanda
        city:
          type: string
          example: Kigali
        address:
          type: string
          example: KK 1 st
        type:
          type: string
          example: 3 bedroom
    body:
      type: object
      properties:
        price:
          type: integer
          example: 10000000
        state:
          type: string
          example: Rwanda
        city:
          type: string
          example: Kigali
        address:
          type: string
          example: KK 1 st
        type:
          type: string
          example: 3 Bed room
        image:
          type: string
          format: binary
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT